Created "create_scale_cylinder"
